+++ CHANGES.rst
@@ -11,6 +11,11 @@ Major changes
   - SciPy (>= 1.2)
   - scikit-learn (>= 0.20.0)
 
+* **SuperVectorizer**: Added automatic transform through the
+  :class:`SuperVectorizer` class. It classifies and transforms
+  columns automatically based on their type. A replacement
+  for the `ColumnTransformer`.
+
 
 Bug-fixes
 ---------